[Reporting] Parse the config object for export csv from dashboard: from query:queryString:options
