+++ x-pack/legacy/plugins/reporting/export_types/csv_from_savedobject/types.d.ts
@@ -146,75 +146,6 @@ export interface SearchSource {
   filter: any[];
 }
 
-export interface SearchRequest {
-  index: string;
-  body:
-    | {
-        _source: {
-          excludes: string[];
-          includes: string[];
-        };
-        docvalue_fields: string[];
-        query:
-          | {
-              bool: {
-                filter: any[];
-                must_not: any[];
-                should: any[];
-                must: any[];
-              };
-            }
-          | any;
-        script_fields: any;
-        sort: Array<{
-          [key: string]: {
-            order: string;
-          };
-        }>;
-        stored_fields: string[];
-      }
-    | any;
-}
-
-export interface SavedSearchGeneratorResult {
-  content: string;
-  maxSizeReached: boolean;
-  size: number;
-}
-
-export interface CsvResultFromSearch {
-  type: string;
-  result: SavedSearchGeneratorResult;
-}
-
-type EndpointCaller = (method: string, params: any) => Promise<any>;
-type FormatsMap = Map<
-  string,
-  {
-    id: string;
-    params: {
-      pattern: string;
-    };
-  }
->;
-
-export interface GenerateCsvParams {
-  searchRequest: SearchRequest;
-  callEndpoint: EndpointCaller;
-  fields: string[];
-  formatsMap: FormatsMap;
-  metaFields: string[]; // FIXME not sure what this is for
-  conflictedTypesFields: string[]; // FIXME not sure what this is for
-  cancellationToken: CancellationToken;
-  settings: {
-    separator: string;
-    quoteValues: boolean;
-    timezone: string | null;
-    maxSizeBytes: number;
-    scroll: { duration: string; size: number };
-  };
-}
-
 /*
  * These filter types are stub types to help ensure things get passed to
  * non-Typescript functions in the right order. An actual structure is not
@@ -236,7 +167,7 @@ export interface SearchSourceFilter extends Filter {
 
 export interface ESQueryConfig {
   allowLeadingWildcards: boolean;
-  queryStringOptions: boolean;
+  queryStringOptions: object;
   ignoreFilterIfFieldNotInIndex: boolean;
 }
